Add option to proxy records synced to Cloudflare.
Previously it was only possible to proxy a record by defining `proxied: true` in the YAML entry of that record. This can't be used when we are syncing between another source than the YAML source. This commit adds the possibility to define this for all records, no matter the source used.